add a require for database connection for more dry code
also added more comments

+++ deletelist.php
@@ -1,22 +1,15 @@
 <?php include 'header.php' ?>
 <?php include 'footer.php' ?>
-<?php 
-$servername = "localhost";
-$username = "root";
-$password = "mysql";
-$myDB = "todolist";
-
-try {
-    $conn = new PDO("mysql:host=$servername;dbname=$myDB", $username, $password);
-    // set the PDO error mode to exception
-    $conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-    //echo "Connected successfully";
-    }
-
-catch(PDOException $e)
-    {
-    echo "Connection failed: " . $e->getMessage();
-    }
+<?php require 'require.php' ?>
+<?php
+/*
+*a file that deletes a list and the tasks that belong to it
+*/
+if (isset($e)) {
+    exit("Unable to connect");
+    header("Location: home.php");   
+}
+else{
     $id= $_GET['id'];
     $stmt = $conn->prepare("DELETE FROM `lists` WHERE id =:id");
     $stmt->bindParam(':id',$id);
@@ -26,3 +19,4 @@ catch(PDOException $e)
     $stmt->bindParam(':listid',$id);
     $stmt->execute();
     header("Location: home.php");
+}
